denote notes directly on wsl, speed fix
- previously the notes were on windows system and it would take time
for them to get fetched, each denote command or simply opening a file
would take longer than usual. This was annoying.
- moved all the note files to wsl. Then created a script that would
copy all the note files into windows machine, onedrive, for them to be
synced with the cloud. This will act as a security measure in case I
forget to commit my notes to git and my machine dies. I will still
have a quite recent copy of the notes in onerive.